+++ LOTTERIA_Kiosk/LOTTERIA_Kiosk/View/CardPayment.xaml.cs
@@ -44,9 +44,9 @@ namespace LOTTERIA_Kiosk.View
                 NavigationService.GoBack();
             }
         }
-        private int GetTotalPrice()
+        private double GetTotalPrice()
         {
-            int total = 0;
+            double total = 0;
             foreach (Food food in App.SelectedMenuList)
             {
                 total += food.Count * food.Price;
